+++ src/main/java/com/magnorion/school/controllers/MonitorController.java
@@ -29,7 +29,7 @@ public class MonitorController {
   }
 
   @RequestMapping(value = "/monitor/{id}/excluir", method = RequestMethod.GET)
-  public String exclusao(@PathVariable Integer id, ModelMap model) {
+  public String exclusao(@PathVariable Long id, ModelMap model) {
     this.service.excluir(id);
 
     return "redirect:/monitor/lista";
