+++ src/main/java/com/magnorion/school/models/domain/Participante.java
@@ -7,18 +7,18 @@ import javax.persistence.*;
 public class Participante {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
-    protected Integer id;
+    protected Long id;
 
     protected String nome;
     protected String telefone;
     protected String email;
     protected Integer idade;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
