+++ src/main/java/com/magnorion/school/repository/AlunoRepository.java
@@ -3,5 +3,5 @@ package com.magnorion.school.repository;
 import com.magnorion.school.models.domain.Aluno;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlunoRepository extends CrudRepository<Aluno, Integer> {
+public interface AlunoRepository extends CrudRepository<Aluno, Long> {
 }
