+++ src/main/java/com/magnorion/school/repository/CursoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CursoRepository extends CrudRepository<Curso, Integer> {
+public interface CursoRepository extends CrudRepository<Curso, Long> {
 }
