+++ src/main/java/com/magnorion/school/repository/DisciplinaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DisciplinaRepository extends CrudRepository<Disciplina, Integer> {
+public interface DisciplinaRepository extends CrudRepository<Disciplina, Long> {
 }
