+++ src/main/java/com/magnorion/school/repository/MonitorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MonitorRepository extends CrudRepository<Monitor, Integer> {
+public interface MonitorRepository extends CrudRepository<Monitor, Long> {
 }
