+++ src/main/java/com/magnorion/school/repository/ProfessorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfessorRepository extends CrudRepository<Professor, Integer> {
+public interface ProfessorRepository extends CrudRepository<Professor, Long> {
 }
